Updates to section 07. Saving sequencer.pio for posterity, going to scrap it
